runtime/evm: move encrypted data stuff
